==> 03_bal-object/03h_export.ipynb <==
import sys
sys.path.insert(0, "../lib")

import math
import os
import pathlib
import re

import scanpy as sc
import numpy as np
import pandas as pd
import sc_utils
import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns

pd.options.display.max_columns = None

%config InlineBackend.figure_format = "retina"

DATA = pathlib.Path('../../data/31_bal-object/')

adata = sc.read_h5ad(DATA / '03_bal-object/03_bal-object.h5ad')

raw_object = sc.read_h5ad('../../data/31_bal-object/raw_object.h5ad')

adata.obs.cell_type = adata.obs.cell_type.astype(str).replace({
    'MoAM-3 mature': 'MoAM-2',
    'MoAM-4 profibrotic': 'MoAM-3 profibrotic',
    'gd/NKT cells': 'NK + γδT cells',
    'Perivascular macrophages': 'Interstitial macrophages',
}).astype('category')

adata.obs.Status = pd.Categorical(adata.obs.Status.astype(str).replace({
    'control': 'Control',
    'SSc': 'SSc-ILD'
}), categories=['Control', 'SSc-ILD'])

paper_ids = pd.read_csv('../SSc_BAL_GEO_ID_Mapping - ID mapping.csv')

adata.obs['paper_id'] = paper_ids.set_index('External Sample ID')['Publication ID'][
    adata.obs['External Sample ID'].values
].values
adata.obs['paper_id'] = pd.Categorical(adata.obs['paper_id'])

raw_object.obs['paper_id'] = paper_ids.set_index('External Sample ID')['Publication ID'][
    raw_object.obs['External Sample ID'].values
].values
raw_object.obs['paper_id'] = pd.Categorical(raw_object.obs['paper_id'])

raw_object.obs_names = (
    raw_object.obs['paper_id'].astype(str)
    + '_'
    + raw_object.obs_names.str.split('-').str[0]
)

adata.obs.drop(columns=[
    'Sample', 'Sample ID', 'External Sample ID',
    'Patient', 'Smoking status', 'Race', 'BMI',
    'Type', 'Tissue location', 'Additional data',
    'Fastq', 'Genome build', 'Directory', 'Doublet threshold',
    'batch', 'n_counts', 'pct_counts_in_top_20_genes',
    'total_counts_mito', 'total_counts_ribo', '_scvi_batch',
    '_scvi_labels', 'leiden_scVI'
], inplace=True)

adata.obs.rename(columns={
    'Study': 'Cohort',
    'Status': 'Condition',
    'paper_id': 'Sample',
}, inplace=True)

adata.obs = adata.obs.loc[
    :,
    [
        'Cohort', 'Sample', 'Condition', 'Sex', 'Age',
        'Protocol', 'Chemistry', 'n_genes_by_counts', 'total_counts', 'pct_counts_in_top_10_genes',
        'pct_counts_mito', 'pct_counts_ribo', 'cell_type'
    ]
]

adata.obs_names = (
    adata.obs.Sample.astype(str)
    + '_'
    + adata.obs_names.str.split('_').str[-1]
)

adata.obs.Cohort = adata.obs.Cohort.astype(str).replace({
    'this': 'Northwestern/Yale'
}).astype('category')

adata.obs.Sex = adata.obs.Sex.str.capitalize()

adata.obs.Protocol = adata.obs.Protocol.astype(str).replace({
    'Wash, no sorting': 'no FACS',
    'FACSorted': 'FACS'
}).astype('category')

age_bins = {}
for i in range(100):
    bin_start = i // 5 * 5
    age_bins[str(i)] = f'{bin_start}–{bin_start + 4}'

adata.obs.Age = adata.obs.Age.astype(str).replace(age_bins).astype('category')

os.makedirs(DATA / '03x_bal-export', exist_ok=True)

First compute markers on log1p-processed `.raw`

sc.tl.rank_genes_groups(adata, "cell_type", method="t-test", n_genes=200)

markers = sc_utils.get_markers(adata, "cell_type")

markers.sort_values(["cluster", "avg_logFC"], ascending=[True, False], inplace=True)

markers.to_csv(DATA / '03x_bal-export/03x_bal-export-markers.csv')

Now subset `raw_object` and put it to `.raw`

current_raw = adata.raw.to_adata()

new_raw = raw_object[current_raw.obs_names].copy()
new_raw.obs = current_raw.obs.copy()

adata.raw = new_raw

adata.write_h5ad(DATA / '03x_bal-export/03x_bal-export.h5ad')
