==> 03_bal-object/05a_gsva_pseudobulk.ipynb <==
import os
import pathlib

import scanpy as sc
import pandas as pd
import numpy as np
import matplotlib as mpl
import matplotlib.pyplot as plt
import decoupler

pd.options.display.max_rows = 300

%config InlineBackend.figure_format = "retina"

# Nick needs this because he's using Firefox which apparently noone uses anymore
from IPython.display import display, HTML
display(HTML("<style>.jp-OutputArea-output {display:flex}</style>"))

def sanitize_name(name):
    return name.replace(' ', '_').replace('*', '').replace(';', '_and').replace('/', '_')

DATA = pathlib.Path('../../data/31_bal-object/')

ds = sc.read_h5ad(DATA / "03_bal-object/03_bal-object.h5ad")

ds.obs.cell_type.unique().tolist()

raw_object = sc.read_h5ad(DATA / "raw_object.h5ad")

raw_object.obs_names = (
    raw_object.obs["External Sample ID"].astype(str) 
    + "_" 
    + raw_object.obs_names.str.replace("(-\d+)+$", "", regex=True)
)

OUT_DIR = DATA / 'pseudobulk-gsva'

CUTOFF = 50
for ct in ds.obs.cell_type.unique():
    ct_slug = sanitize_name(ct)
    cells = ds.obs.index[ds.obs.cell_type.eq(ct)]
    patient_status = ds.obs.loc[
        cells, 
        ["External Sample ID", "Status", "Sex", "Study", 'Chemistry']
    ].drop_duplicates()
    samples = patient_status["External Sample ID"].unique()
    sample_values = [] # gene sum for sample in this cell type
    sample_ncells = [] # number of cells expressing gene for sample in this cell type
    filtered_samples = [] # samples that pass cutoff filter
    n_cells = []
    for s in samples:
        # cells in particular sample
        s_cells = raw_object.obs_names.isin(cells) & raw_object.obs["External Sample ID"].eq(s)
        if s_cells.sum() >= CUTOFF:
            sample_values.append(raw_object.X[s_cells, :].sum(axis=0).A[0])
            sample_ncells.append((raw_object.X[s_cells, :] > 0).sum(axis=0).A[0])
            filtered_samples.append(s)
            n_cells.append(s_cells.sum())
            
    sample_values = pd.DataFrame(sample_values, index=filtered_samples, columns=raw_object.var_names).T
    sample_ncells = pd.DataFrame(sample_ncells, index=filtered_samples, columns=raw_object.var_names).T

    os.makedirs(f'{OUT_DIR}/{ct_slug}/data', exist_ok=True)
    
    fname = f"{OUT_DIR}/{ct_slug}/data/{ct_slug}.txt"
    sample_values.to_csv(fname, sep="\t")
    
    fname = f"{OUT_DIR}/{ct_slug}/data/{ct_slug}-n_cells.txt"
    sample_ncells.to_csv(fname, sep="\t")
    
    fname = f"{OUT_DIR}/{ct_slug}/data/{ct_slug}-meta.csv"
    patient_status = patient_status.loc[patient_status["External Sample ID"].isin(filtered_samples), :]
    patient_status["n_cells"] = n_cells
    patient_status.to_csv(fname)
    print(f"{ct} done, {len(filtered_samples)}/{len(samples)}")

ct_slug = 'global'
cells = ds.obs.index
patient_status = ds.obs.loc[
    cells, 
    ["External Sample ID", "Status", "Sex", "Study", 'Chemistry']
].drop_duplicates()
samples = patient_status["External Sample ID"].unique()
sample_values = [] # gene sum for sample in this cell type
sample_ncells = [] # number of cells expressing gene for sample in this cell type
filtered_samples = [] # samples that pass cutoff filter
n_cells = []
for s in samples:
    # cells in particular sample
    s_cells = raw_object.obs_names.isin(cells) & raw_object.obs["External Sample ID"].eq(s)
    if s_cells.sum() >= CUTOFF:
        sample_values.append(raw_object.X[s_cells, :].sum(axis=0).A[0])
        sample_ncells.append((raw_object.X[s_cells, :] > 0).sum(axis=0).A[0])
        filtered_samples.append(s)
        n_cells.append(s_cells.sum())
        
sample_values = pd.DataFrame(sample_values, index=filtered_samples, columns=raw_object.var_names).T
sample_ncells = pd.DataFrame(sample_ncells, index=filtered_samples, columns=raw_object.var_names).T

os.makedirs(f'{OUT_DIR}/{ct_slug}/data', exist_ok=True)

fname = f"{OUT_DIR}/{ct_slug}/data/{ct_slug}.txt"
sample_values.to_csv(fname, sep="\t")

fname = f"{OUT_DIR}/{ct_slug}/data/{ct_slug}-n_cells.txt"
sample_ncells.to_csv(fname, sep="\t")

fname = f"{OUT_DIR}/{ct_slug}/data/{ct_slug}-meta.csv"
patient_status = patient_status.loc[patient_status["External Sample ID"].isin(filtered_samples), :]
patient_status["n_cells"] = n_cells
patient_status.to_csv(fname)
print(f"{ct_slug} done, {len(filtered_samples)}/{len(samples)}")
